requirements.txt
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.115.14
greenlet==3.2.3
idna==3.10
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
